Add files via upload
Here is an updated version of the rig with the empties preloaded.  Each of the bones has named constraints that should adequately describe what they are doing.  I mostly addressed a parenting loop caused by the right hip on the "Hip Direction For Pole Targets" object, which caused the kneed poles to freak out.  I also added another joint to the torso, and bones on the legs that can be used with drivers to get local location from the feet for inverse kinematics.